Refactor FirebaseAuth, Datacollection, FirebaseCollection and CurrentUserManager class
