+++ app/src/main/java/com/andela/helpmebuy/dal/DataCollection.java
@@ -1,7 +1,6 @@
 package com.andela.helpmebuy.dal;
 
 
-import java.util.LinkedHashMap;
 import java.util.List;
 
 public interface DataCollection<T> {
@@ -13,6 +12,4 @@ public interface DataCollection<T> {
     void getAll(DataCallback<List<T>> callback);
 
     void query(String path, String arg, DataCallback<List<T>> callback);
-
-    void update();
 }
